Move driver code to driver.c file - move code outside of retroarch.c
