+++ Makefile.common
@@ -234,6 +234,7 @@ endif
 
 OBJ += frontend/frontend_driver.o \
        retroarch.o \
+       driver.o \
        camera/camera_driver.o \
        record/record_driver.o \
        command.o \
